+++ src/env_sensor.h
@@ -3,8 +3,8 @@
 namespace env_sensor
 {
 
-extern double temperature;
-extern double humidity;
+extern float temperature;
+extern float humidity;
 
 void setup(uint8_t pin);
 void loop();
